Basic Dockerfile and Compose Setup as a Starting Point
After some experimentation, I finally managed to build an image that stays running instead of crashing on startup. I used ubuntu:24.04 as the base image — definitely heavier than ideal, but it worked reliably — and set rethink-cloud.js as the container’s entry point.
I'm still not entirely sure which files need to persist, but for now I’ve mounted only the configuration file (/rethink/config.json) into the container. With that in place, the output behaves exactly as described in the wiki.

In my case i created a DNS Record for both common.lgthinq.com and lgthinq.mydomain.tld pointing to my server with NginX
There I redirected one domain to the other.
Is it posible in this case to simply use port 443 and let NginX handle the certificates ?

I have not yet tried it with my ThinQ machine because I do not have a wifi enabled device where i can run rethink-setup.js

This should be the output if rethink-cloud.js was correctly configured and mapped to the container

root@homeserver:/services/rethink# docker compose logs -f rethink
rethink  | Rethink cloud ready
rethink  | During setup, please ensure that connections to common.lgthinq.com:443 are redirected to lgthinq.mydomain.tld:443
rethink  | HA mqtt connection established

This byfar not a production ready solution but it is just a start to get things working